:sparkles: Add `metrics` timeseries
This timeseries will have both `std_dev` and `returns` observations

:bug: Validate `returns` and `std_dev` timeseries

:bug:  Sort the index of price series

This fixes an issue where the std_dev ts would be incorrect,
since the rolling window calcs are applied on a decending ordered ts.

This means that the last seven days would not have std_dev observations,
but the first seven would.

:rewind: Reintroduce asset property for frontend